services/ui_parser/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
Pillow>=10.0.0
pytesseract>=0.3.10
# GroundingDINO (visual element detection)
torch>=2.1.0
transformers>=4.50.0
scipy>=1.10.0
